requirements.txt
ahk==1.1.2
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
ci-info==0.3.0
click==8.1.3
clr-loader==0.2.4
colorama==0.4.5
comtypes==1.2.0
configobj==5.0.8
configparser==5.3.0
cryptography==40.0.2
et-xmlfile==1.1.0
etelemetry==0.3.0
fastapi==0.85.1
filelock==3.12.2
fitz==0.0.1.dev2
Flask==2.3.2
frozenlist==1.3.3
future==0.18.3
grafanalib==0.7.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
looseversion==1.3.0
lxml==4.9.2
MarkupSafe==2.1.2
MouseInfo==0.1.3
multidict==6.0.4
multipledispatch==1.0.0
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.23.4
opcua==0.98.13
openai==0.27.6
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==23.1
pandas==1.5.1
pathlib==1.0.1
pefile==2023.2.7
Pillow==9.4.0
prov==2.0.0
proxy-tools==0.1.0
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==1.10.2
PyDirectInput==1.0.4
pydot==1.4.2
pygame==2.1.2
PyGetWindow==0.0.9
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyMsgBox==1.0.9
PyMuPDF==1.22.5
pyparsing==3.1.0
PyPDF2==3.0.1
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
PyQt5Designer==5.14.1
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.10
python-dateutil==2.8.2
pythonnet==3.0.1
pytweening==1.0.4
pytz==2022.6
pywebview==3.7.2
pywin32==304
pywin32-ctypes==0.2.0
pywinauto==0.6.8
pyxnat==1.5
rdflib==6.3.2
requests==2.28.1
scipy==1.11.1
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.20.4
tqdm==4.65.0
traits==6.3.2
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.19.0
Werkzeug==2.3.4
xlrd==2.0.1
XlsxWriter==3.0.3
xlwings==0.28.5
xlwt==1.3.0
yarl==1.9.2
